+++ oak_functions/loader/benches/lookup.rs
@@ -25,7 +25,7 @@ use criterion::{
 };
 use hashbrown::HashMap;
 use lookup_data_generator::data::generate_and_serialize_random_entries;
-use oak_functions_abi::proto::{Request, StatusCode};
+use oak_functions_abi::{proto::StatusCode, Request};
 use oak_functions_loader::{
     logger::Logger, lookup_data::parse_lookup_entries, server::WasmHandler,
 };
@@ -162,7 +162,7 @@ fn run_lookup_iteration(
         body: benchmark_request.to_owned(),
     };
     let resp = wasm_handler.handle_invoke(request).unwrap();
-    assert_eq!(resp.status, StatusCode::Success as i32);
+    assert_eq!(resp.status, StatusCode::Success);
     assert_eq!(&resp.body, expected_value);
 }
